fxinst: use native GSU data types (mostly) - Most commonly-used values in the FxRegs_s struct now use GSU-native data types, with the exceptions being avReg as those will require more effort to convert, and the status register optimizations stilll use 32-bit numbers. - Reordered the fields of the FxRegs_s struct to be slightly more optimal. Some rarely-used values have been pushed further into the struct. - Replaced vOpcode and vLow with native data types. I'm not sure why, but this was necessary to prevent extra instructions from generating after vStatusReg was reduced to a u16. title 20.41 -> 19.84 1-3 pipe 5.50 -> 5.50 1-3 cave 11.20 -> 10.95

+++ source/Snes9x/fxinst.cpp
@@ -1237,8 +1237,8 @@ static uint32 fx_run(uint32 nInstructions)
     READR14;
     while(LIKELY(vCounter-- > 0))
     {
-        uint32 vOpcode = (uint32)PIPE | (GSU.vStatusReg & (FLG_ALT1 | FLG_ALT2));
-        uint32 vLow = vOpcode & 0xf;
+        uint16 vOpcode = PIPE | (GSU.vStatusReg & (FLG_ALT1 | FLG_ALT2));
+        uint8 vLow = vOpcode & 0xf;
         FETCHPIPE;
 
         ASSUME(vOpcode < 0x400);
